Harden protected routes and split app shell client
